LDM with VQVAE/requirements.txt
boto3
botocore
certifi
charset-normalizer
click
contourpy
cycler
einops
filelock
fonttools
fsspec
ftfy
huggingface-hub
idna
importlib-metadata
importlib-resources
jmespath
joblib
kiwisolver
matplotlib
numpy
opencv-python
packaging
pandas
Pillow
pyparsing
python-dateutil
PyYAML
regex
requests
s3transfer
sacremoses
safetensors
scipy
sentencepiece
six
tokenizers
torch
torchvision
tqdm
transformers
typing_extensions
urllib3
wcwidth
zipp
